refactor : 도메인간 결합도 낮추기 위한 리팩터링  - 도메인은 다른 도메인을 모르게끔 Facade에서 조율한다.  - "카트"에 "상품"을 담을 때에는 어떻게 처리해야할 지 전혀 감이 오지않아 일단 서비스까지는 Product를 넘겨서 처리. 이후 질의 필요

+++ src/main/java/com/hanghae/ecommerce/domain/product/ProductInfo.java
@@ -4,7 +4,7 @@ public record ProductInfo(
 	Long productId,
 	String productName,
 	Long productPrice,
-	int productStock
+	Long productStock
 ) {
 	public static ProductInfo from(Product product) {
 		return new ProductInfo(
